requirements.txt
dash
scikit_image
plotly
dash_bootstrap_components
dataclasses
mashumaro
pandas
Pillow
scikit-image
pyyaml

opencv-python
python-dotenv
tqdm
torch
torchvision
git+https://github.com/facebookresearch/segment-anything.git